Change AZRandomString to category

+++ idap/CarWash/Cars/AZCar.m
@@ -7,8 +7,8 @@
 //
 
 #import "AZCar.h"
-#import "AZRandomString.h"
 #import "AZWasher.h"
+#import "NSNumber+AZRandomNumber.h"
 
 static NSString * const AZDescriptionFormatter = @"%@: %@";
 
@@ -28,6 +28,7 @@ const u_int AZMaxMoneyForCar = 5000;
     
     [super dealloc];
 }
+
 
 - (instancetype)init {
     [super init];
@@ -66,10 +67,10 @@ const u_int AZMaxMoneyForCar = 5000;
 #pragma mark -
 #pragma mark Implements protocols
 
-- (double)giveMoney:(AZWasher *)washer {
-    double result = self.money;
+- (NSUInteger)giveMoney:(AZWasher *)washer {
+    NSUInteger result = self.money;
     self.money = 0;
-    NSLog(@"%@ give %5.2f dollars to %@ ", self, result, washer);
+    NSLog(@"%@ give %lu dollars to %@ ", self, result, washer);
     
     return result;
 }
